+++ packages/common/web/src/Utils/viewerUtils.tsx
@@ -5,6 +5,8 @@ import {
   RICOS_TEXT_COLOR_TYPE,
   RICOS_TEXT_HIGHLIGHT_TYPE,
   RICOS_FONT_SIZE_TYPE,
+  RICOS_BOLD_TYPE,
+  RICOS_ITALIC_TYPE,
   TEXT_HIGHLIGHT_TYPE,
   TEXT_COLOR_TYPE,
   DraftContent,
@@ -26,7 +28,10 @@ const RICOS_TO_DRAFT_TYPE = {
   [RICOS_TEXT_HIGHLIGHT_TYPE]: TEXT_HIGHLIGHT_TYPE,
 };
 
-export type CustomInlineStyleType = keyof typeof INLINE_STYLE_TYPES;
+export type CustomInlineStyleType =
+  | keyof typeof INLINE_STYLE_TYPES
+  | typeof RICOS_BOLD_TYPE
+  | typeof RICOS_ITALIC_TYPE;
 
 export const parseStyleByType = (style: string, type: CustomInlineStyleType) => {
   const property = INLINE_STYLE_TYPES[type];
@@ -57,6 +62,13 @@ export const dynamicStyleParsers = {
   [RICOS_FONT_SIZE_TYPE]: (style: string) => parseStyleByType(style, RICOS_FONT_SIZE_TYPE),
 };
 
+export const draftDecorationsToCss = {
+  BOLD: { 'font-weight': 'bold' },
+  ITALIC: { 'font-style': 'italic' },
+  NOT_BOLD: { 'font-weight': 'normal' },
+  NOT_ITALIC: { 'font-style': 'normal' },
+};
+
 export const defaultStyleFnMapper = {
   [RICOS_TEXT_COLOR_TYPE]: (style: string) => {
     const parsedStyle = dynamicStyleParsers[RICOS_TEXT_COLOR_TYPE](style, isHexColor);
@@ -70,6 +82,10 @@ export const defaultStyleFnMapper = {
     const parsedStyle = dynamicStyleParsers[RICOS_FONT_SIZE_TYPE](style);
     return parsedStyle ? { fontSize: parsedStyle } : undefined;
   },
+  [RICOS_BOLD_TYPE]: (style: string) =>
+    style === 'NOT_BOLD' ? { fontWeight: 'normal' } : undefined,
+  [RICOS_ITALIC_TYPE]: (style: string) =>
+    style === 'NOT_ITALIC' ? { fontStyle: 'normal' } : undefined,
 };
 
 const defaultStyleSelectionPredicates = {