+++ packages/editor/web/src/RichContentEditor/EditorCommands.ts
@@ -12,7 +12,6 @@ import {
   redo,
   getTextAlignment,
   hasInlineStyle,
-  getDraftInlineStyle,
   getBlockType,
   hasLinksInSelection,
   getLinkDataInSelection,
@@ -25,15 +24,13 @@ import {
   mergeBlockData,
   getAnchorBlockData,
   getAnchorableBlocks,
-  setFontSize,
-  getFontSize,
-  getColor,
-  setHighlightColor,
-  setTextColor,
+  removeCurrentInlineStyle,
   isAtomicBlockInSelection,
+  scrollToBlock,
   insertCustomLink,
 } from 'wix-rich-content-editor-common';
 import {
+  AvailableExperiments,
   EditorCommands,
   GetEditorState,
   SetEditorState,
@@ -99,9 +96,24 @@ import {
   RICOS_CODE_BLOCK_TYPE,
   RICOS_FONT_SIZE_TYPE,
   UNSUPPORTED_BLOCKS_TYPE,
+  DocumentStyle,
+  RicosCustomStyles,
   EXTERNAL,
   CUSTOM_LINK,
 } from 'wix-rich-content-common';
+
+import {
+  getDocumentStyle,
+  getWiredFontStyles,
+  updateDocumentStyle,
+  getAnchorBlockInlineStyles,
+  setFontSize,
+  getFontSize,
+  getColor,
+  setHighlightColor,
+  setTextColor,
+  toggleInlineStyle,
+} from './utils/editorCommandsUtils';
 
 const TO_DRAFT_PLUGIN_TYPE_MAP = {
   [RICOS_DIVIDER_TYPE]: DIVIDER_TYPE,
@@ -207,7 +219,9 @@ export const createEditorCommands = (
   createPluginsDataMap,
   plugins,
   getEditorState: GetEditorState,
-  setEditorState: SetEditorState
+  setEditorState: SetEditorState,
+  externalEditorProps,
+  experiments?: AvailableExperiments
 ): EditorCommands => {
   const setBlockType: EditorCommands['setBlockType'] = type => {
     setEditorState(RichUtils.toggleBlockType(getEditorState(), type));
@@ -220,6 +234,10 @@ export const createEditorCommands = (
         SelectionState.createEmpty(blockKey).merge(selection || {})
       )
     );
+
+  const documentStyleGetter = externalEditorProps.getDocumentStyle
+    ? externalEditorProps.getDocumentStyle
+    : () => getDocumentStyle(getEditorState());
 
   const editorState: {
     getSelection: EditorCommands['getSelection'];
@@ -235,16 +253,27 @@ export const createEditorCommands = (
     getLinkDataInSelection: EditorCommands['getLinkDataInSelection'];
     getSelectedData: EditorCommands['getSelectedData'];
     getPluginsList: EditorCommands['getPluginsList'];
+    scrollToBlock: EditorCommands['scrollToBlock'];
+    isBlockInContent: EditorCommands['isBlockInContent'];
     getBlockSpacing: EditorCommands['getBlockSpacing'];
     saveEditorState: EditorCommands['saveEditorState'];
     loadEditorState: EditorCommands['loadEditorState'];
     saveSelectionState: EditorCommands['saveSelectionState'];
     loadSelectionState: EditorCommands['loadSelectionState'];
+    getDocumentStyle: EditorCommands['getDocumentStyle'];
+    updateDocumentStyle: EditorCommands['updateDocumentStyle'];
+    getAnchorBlockInlineStyles: EditorCommands['getAnchorBlockInlineStyles'];
+    getWiredFontStyles: EditorCommands['getWiredFontStyles'];
     isAtomicBlockInSelection: EditorCommands['isAtomicBlockInSelection'];
   } = {
     getSelection: () => {
       const selection = getEditorState().getSelection();
-      return { getIsCollapsed: selection.isCollapsed(), getIsFocused: selection.getHasFocus() };
+      return {
+        isCollapsed: selection.isCollapsed(),
+        isFocused: selection.getHasFocus(),
+        startKey: selection.getStartKey(),
+        endKey: selection.getEndKey(),
+      };
     },
     getAnchorableBlocks: () => getAnchorableBlocks(getEditorState()),
     getColor: colorType => getColor(getEditorState(), colorType),
@@ -278,7 +307,34 @@ export const createEditorCommands = (
         (pluginName: string) => pluginName && !PluginsToExclude.includes[pluginName]
       );
     },
+    getDocumentStyle: () => documentStyleGetter(),
+    updateDocumentStyle: (documentStyle: DocumentStyle) => {
+      if (externalEditorProps.updateDocumentStyle) {
+        externalEditorProps.updateDocumentStyle(documentStyle);
+      } else {
+        const newEditorState = updateDocumentStyle(getEditorState(), documentStyle);
+        setEditorState(newEditorState);
+      }
+    },
+    getAnchorBlockInlineStyles: () => getAnchorBlockInlineStyles(getEditorState()),
+    getWiredFontStyles: (customStyles?: RicosCustomStyles, isMobile?: boolean) =>
+      getWiredFontStyles(documentStyleGetter(), customStyles, isMobile),
+    scrollToBlock: blockKey => scrollToBlock(blockKey, experiments),
+    isBlockInContent: blockKey => {
+      const blocks = getEditorState()
+        .getCurrentContent()
+        .getBlocksAsArray();
+      return blocks.some(block => block.getKey() === blockKey);
+    },
     isAtomicBlockInSelection: () => isAtomicBlockInSelection(getEditorState()),
+  };
+
+  const toggleOverlayBGColor = (element: HTMLElement) => {
+    if (!element.style.backgroundColor) {
+      element.style.backgroundColor = 'rgba(var(--ricos-text-color-tuple, 33, 33, 33), 0.1)';
+    } else {
+      element.style.backgroundColor = '';
+    }
   };
 
   const textFormattingCommands: {
@@ -288,17 +344,24 @@ export const createEditorCommands = (
     setBlockType: EditorCommands['setBlockType'];
     setTextAlignment: EditorCommands['setTextAlignment'];
     _setSelection: EditorCommands['_setSelection'];
+    toggleBlockOverlay: EditorCommands['toggleBlockOverlay'];
   } = {
     undo: () => setEditorState(undo(getEditorState())),
     redo: () => setEditorState(redo(getEditorState())),
     toggleInlineStyle: inlineStyle =>
-      setEditorState(
-        RichUtils.toggleInlineStyle(getEditorState(), getDraftInlineStyle(inlineStyle))
-      ),
+      setEditorState(toggleInlineStyle(getEditorState(), inlineStyle)),
     setBlockType,
     setTextAlignment: textAlignment =>
       setEditorState(setTextAlignment(getEditorState(), textAlignment)),
     _setSelection,
+    toggleBlockOverlay: blockKey => {
+      const blockElement = document.querySelector(`[data-offset-key="${blockKey}-0-0"]`);
+      const elementOverlay = blockElement?.querySelector(`[data-hook="componentOverlay"]`);
+      const element = elementOverlay
+        ? blockElement?.querySelector(`[data-hook="componentOverlay"]`)
+        : blockElement;
+      toggleOverlayBGColor(element as HTMLElement);
+    },
   };
 
   const pluginsCommands: {
@@ -334,6 +397,7 @@ export const createEditorCommands = (
     insertDecoration: EditorCommands['insertDecoration'];
     triggerDecoration: EditorCommands['triggerDecoration'];
     deleteDecoration: EditorCommands['deleteDecoration'];
+    clearSelectedBlocksInlineStyles: EditorCommands['clearSelectedBlocksInlineStyles'];
   } = {
     insertDecoration: (type: string, data, settings) => {
       const draftType = TO_DRAFT_PLUGIN_TYPE_MAP[type];
@@ -356,6 +420,10 @@ export const createEditorCommands = (
         setEditorState(newEditorState);
       }
     },
+    clearSelectedBlocksInlineStyles: (exclude?: string[]) => {
+      const styleSelectionPredicate = style => !exclude?.includes(style);
+      setEditorState(removeCurrentInlineStyle(getEditorState(), styleSelectionPredicate));
+    },
   };
 
   const editorCommands = {
@@ -364,5 +432,6 @@ export const createEditorCommands = (
     ...decorationsCommands,
     ...editorState,
   };
+
   return editorCommands;
 };