+++ packages/tiptap-editor/web/src/components/RichContentAdapter/RichContentAdapter.tsx
@@ -99,8 +99,8 @@ export class RichContentAdapter implements TiptapAPI {
         return this.editor.commands.updateNodeById(blockKey, data);
       },
       getSelection: () => ({
-        getIsFocused: this.editor.isFocused,
-        getIsCollapsed: this.editor.state.selection.empty,
+        isFocused: this.editor.isFocused,
+        isCollapsed: this.editor.state.selection.empty,
       }),
 
       insertDecoration: (type, data) => {
@@ -217,6 +217,9 @@ export class RichContentAdapter implements TiptapAPI {
     isRedoStackEmpty: () => false,
     getSelectedData: () => 'blah',
     getPluginsList: () => [],
+    scrollToBlock: _blockKey => {},
+    isBlockInContent: _blockKey => false,
+    toggleBlockOverlay: _blockKey => false,
     getBlockSpacing: () => 5,
     saveEditorState: () => {},
     loadEditorState: () => {},
@@ -228,6 +231,13 @@ export class RichContentAdapter implements TiptapAPI {
     setBlockType: () => {},
     setTextAlignment: () => {},
     _setSelection: () => {},
+    getDocumentStyle: () => undefined,
+    getAnchorBlockInlineStyles: () => {
+      return {};
+    },
+    updateDocumentStyle: () => {},
+    clearSelectedBlocksInlineStyles: () => {},
+    getWiredFontStyles: () => undefined,
     isAtomicBlockInSelection: () => false,
   };
 }